pyproject.toml
[project]
name = "zendesk-mcp-server"
version = "0.1.0"
description = "A simple Zendesk MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "mcp>=1.1.2",
    "python-dotenv>=1.0.1",
    "zenpy>=2.0.56",
    "PyJWT>=2.8.0",
    "starlette>=0.37.0",
    "cachetools>=5.3.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"
